Adds controller and action to the context
